% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series.R
\name{fredr_series}
\alias{fredr_series}
\alias{fred_series}
\alias{fredr_series_categories}
\alias{fred_series}
\alias{fredr_series_release}
\alias{fred_series}
\alias{fredr_series_tags}
\alias{fred_series}
\alias{fredr_series_updates}
\alias{fred_series}
\alias{fredr_series_vintagedates}
\title{Get information about a FRED series}
\usage{
fredr_series(series_id = NULL, realtime_start = NULL, realtime_end = NULL)

fredr_series_categories(series_id = NULL, realtime_start = NULL,
  realtime_end = NULL)

fredr_series_release(series_id = NULL, realtime_start = NULL,
  realtime_end = NULL)

fredr_series_tags(series_id = NULL, order_by = NULL, sort_order = NULL,
  realtime_start = NULL, realtime_end = NULL)

fredr_series_updates(limit = NULL, offset = NULL, filter_value = NULL,
  start_time = NULL, end_time = NULL, realtime_start = NULL,
  realtime_end = NULL)

fredr_series_vintagedates(series_id = NULL, limit = NULL, offset = NULL,
  sort_order = NULL, realtime_start = NULL, realtime_end = NULL)
}
\arguments{
\item{series_id}{A string ID for the FRED series.}

\item{realtime_start}{A \code{Date} indicating the start of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{realtime_end}{A \code{Date} indicating the end of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{order_by}{A string indicating the attribute by which to order the
Possible values include \code{"series_count"} (default), \code{"popularity"},
\code{"created"}, \code{"name"}, and \code{"group_id"}.}

\item{sort_order}{A string representing the order of the resulting series.
Possible values are: \code{"asc"} (default), and \code{"desc"}.}

\item{limit}{An integer limit on the maximum number of results to return.
Defaults to \code{1000}, the maximum.}

\item{offset}{An integer used in conjunction with \code{limit} for long series.
This mimics the idea of \emph{pagination} to retrieve large amounts of data over
multiple calls. Defaults to \code{0}.}

\item{filter_value}{Filter results by type of geographic region of economic
the data series.  Possible values include
\itemize{
\item \code{"all"} (default) - no filtering
\item \code{"macro"} - filters results macroeconomic regions (e.g. entire countries)
\item \code{"regional"} - filters results to series for regions of the United States
such as states, counties, and Metropolitan Statistical Areas (MSA).
}}

\item{start_time}{A datetime object indicating the start time to filter series
updates results.}

\item{end_time}{A datetime object indicating the start time to filter series
updates results.}
}
\value{
A \code{tibble} object.
}
\description{
Given a series ID, return information on a series as a \code{tibble} object.  See
Details section for a short description of each function.
}
\details{
\itemize{
\item \code{fredr_series()} - Return basic details for a series.  Note that this
function will \emph{not} return the actual series data.  For this functionality,
see \code{\link[=fredr_series_observations]{fredr_series_observations()}}.
\item \code{fredr_series_categories()} - Return the categories for a series.
\item \code{fredr_series_release()} - Return the release for a series.
\item \code{fredr_series_tags()} - Return the tags for a series.
\item \code{fredr_series_updates()} - Return a set of series sorted by when
observations were updated on the FRED server.
\item \code{fredr_series_vintagedates()} - Return a sequence of dates in history
when a series' data values were revised or new data values were released.
}
}
\section{API Documentation}{

\itemize{
\item \href{https://research.stlouisfed.org/docs/api/fred/series.html}{fred/series}
\item \href{https://research.stlouisfed.org/docs/api/fred/series_categories.html}{fred/series/categories}
\item \href{https://research.stlouisfed.org/docs/api/fred/series_release.html}{fred/series/release}
\item \href{https://research.stlouisfed.org/docs/api/fred/series_tags.html}{fred/series/tags}
\item \href{https://research.stlouisfed.org/docs/api/fred/series_updates.html}{fred/series/updates}
\item \href{https://research.stlouisfed.org/docs/api/fred/series_vintagedates.html}{fred/series/vintagedates}
}
}

\examples{
\dontrun{
fredr_series("UNRATE")
fredr_series_categories("UNRATE")
fredr_series_release("UNRATE")
fredr_series_tags("UNRATE", order_by = "group_id")
fredr_series_update(filter_value = "regional")
fredr_series_vintagedates(filter_value = "regional")
}
}
\seealso{
\code{\link[=fredr_series_observations]{fredr_series_observations()}}, \code{\link[=fredr_series_search_text]{fredr_series_search_text()}},
\code{\link[=fredr_series_search_id]{fredr_series_search_id()}}, \code{\link[=fredr_series_search_tags]{fredr_series_search_tags()}},
\code{\link[=fredr_series_search_related_tags]{fredr_series_search_related_tags()}}.
}
