% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr-package.R
\docType{package}
\name{fredr-package}
\alias{fredr-package}
\alias{_PACKAGE}
\title{fredr: An R Client for the 'FRED' API}
\description{
An R client for the 'Federal Reserve Economic Data'
    ('FRED') API <https://api.stlouisfed.org>.  Functions to retrieve
    economic time series and other data from 'FRED'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sboysel/fredr}
  \item Report bugs at \url{https://github.com/sboysel/fredr/issues}
}

}
\author{
\strong{Maintainer}: Sam Boysel \email{sboysel@gmail.com}

Authors:
\itemize{
  \item Davis Vaughan \email{davis@rstudio.com}
}

}
\keyword{internal}
