% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_docs.R
\name{fredr_docs}
\alias{fredr_docs}
\title{Open the web documentation for a certain FRED API topic.}
\usage{
fredr_docs()
}
\description{
Opens FRED API web documentation in a new web browser tab.
}
\section{API Documentation}{


\href{https://api.stlouisfed.org/docs/fred/}{FRED API}
}

\examples{
fredr_docs()
}
