% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_endpoints.R
\docType{data}
\name{fredr_endpoints}
\alias{fredr_endpoints}
\title{List of available FRED API endpoints.}
\format{
A \code{tibble} with 31 rows and 3 variables:
\describe{
\item{endpoint}{endpoint name (e.g. "fred/category", "fredr/series/observations", "fredr/tags").
This name can be supplied to the \code{endpoint} parameter in \code{\link[=fredr_docs]{fredr_docs()}} to
open the \href{https://fred.stlouisfed.org/docs/api/fred/}{FRED API endpoint documentation}
in a web browser.}
\item{type}{endpoint type (e.g. "Categories", "Releases", "Series", "Sources", and "Tags".)}
\item{note}{endpoint details}
}
}
\usage{
fredr_endpoints
}
\description{
List of available FRED API endpoints.
}
\section{API Documentation}{


\href{https://fred.stlouisfed.org/docs/api/fred/}{FRED API}
}

\seealso{
\code{\link[=fredr_request]{fredr_request()}}, \code{\link[=fredr_docs]{fredr_docs()}}
}
\keyword{datasets}
