% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_options.R
\name{fredr_options}
\alias{fredr_options}
\title{Optional parameters used in FRED API calls}
\usage{
fredr_options(observation_start = NULL, observation_end = NULL,
  frequency = NULL, aggregation_method = NULL, limit = NULL,
  offset = NULL, sort_order = NULL, units = NULL, realtime_start = NULL,
  realtime_end = NULL, vintage_dates = NULL, output_type = NULL,
  tag_names = NULL, exclude_tag_names = NULL, tag_group_id = NULL,
  search_text = NULL)
}
\arguments{
\item{observation_start}{A \code{Date} indicating the start of the observation period.
Defaults to \code{1776-07-04}, the earliest available date.}

\item{observation_end}{A \code{Date} indicating the end of the observation period.
Defaults to \code{9999-12-31}, the latest available date.}

\item{frequency}{A string representing a lower frequency to aggregate to.
Defaults to no frequency aggregation. Possible values are:
\itemize{
\item \code{"d"} - Daily
\item \code{"w"} - Weekly
\item \code{"bw"} - Biweekly
\item \code{"m"} - Monthly
\item \code{"q"} - Quarterly
\item \code{"sa"} - Semiannual
\item \code{"a"} - Annual
\item \code{"wem"} - Weekly, ending Monday
\item \code{"wetu"} - Weekly, ending Tuesday
\item \code{"wew"} - Weekly, ending Wednesday
\item \code{"weth"} - Weekly, ending Thursday
\item \code{"wef"} - Weekly, ending Friday
\item \code{"wesa"} - Weekly, ending Saturday
\item \code{"wesu"} - Weekly, ending Sunday
\item \code{"bwew"} - Biweekly, ending Wednesday
\item \code{"bwem"} - Biweekly, ending Monday
}}

\item{aggregation_method}{A string representing the aggregation method
used for frequency aggregation. This parameter has no affect is \code{frequency}
is not set. Possible values are:
\itemize{
\item \code{"avg"} for average
\item \code{"sum"} for sum
\item \code{"eop"} for end of period value
}}

\item{limit}{An integer limit on the maximum number of results to return.
Defaults to the maximum value for the endpoint. Consult the API documentation
of your endpoint for specific information.}

\item{offset}{An integer used in conjunction with \code{limit} for long series.
This mimics the idea of \emph{pagination} to retrieve large amounts of data over
multiple calls. Defaults to \code{0}.}

\item{sort_order}{A string representing the order of the resulting series.
Possible values are: \code{"asc"}, and \code{"desc"}. Generally defaults to \code{"asc"}.}

\item{units}{A string indicating the data value transformation.
Defaults to \code{"lin"}. Possible values are:
\itemize{
\item \code{"lin"} - Levels (No transformation)
\item \code{"chg"} - Change
\item \code{"ch1"} - Change from 1 year ago
\item \code{"pch"} - Percent change
\item \code{"pc1"} - Percent change from 1 year ago
\item \code{"pca"} - Counpounded annual rate of change
\item \code{"cch"} - Continuously compounded rate of change
\item \code{"cca"} - Continuously compounded annual rate of change
\item \code{"log"} - Natural log
}}

\item{realtime_start}{A \code{Date} indicating the start of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{realtime_end}{A \code{Date} indicating the end of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{vintage_dates}{A vector of \code{Date} objects to download data for.
Vintage dates are used to download data as it existed on these specified
dates in history. They can be specified instead of a real-time period using
\code{realtime_start} and \code{realtime_end}. Defaults to no vintage dates.}

\item{output_type}{An integer indicating the output type. Defaults to \code{1}
but is not used unless \code{realtime_start} is used. Possible values
are:
\itemize{
\item \code{1} for Observations by Real-Time Period
\item \code{2} for Observations by Vintage Date, All Observations
\item \code{3} for Observations by Vintage Date, New and Revised Observations Only
\item \code{4} for Observations, Initial Release Only.
}}

\item{tag_names}{A semicolon delimited string of tag names.
Note that this value must be a valid FRED series tag
(e.g. \code{"30-year"}, \code{"usa"}, \code{"quarterly"}, etc.).}

\item{exclude_tag_names}{A semicolon delimited string of tag names used to
exclude results with.}

\item{tag_group_id}{A tag group id to filter tags by type. No filtering by
tag group by default. Possible values are:
\itemize{
\item \code{freq} = Frequency
\item \code{gen} = General or Concept
\item \code{geo} = Geography
\item \code{geot} = Geography Type
\item \code{rls} = Release
\item \code{seas} = Seasonal Adjustment
\item \code{src} = Source
}}

\item{search_text}{A string containing the words to match against.}

\item{order_by}{A string indicating the attribute to order results by.
Consult the API documentation of your endpoint for specific information.}
}
\description{
This helper function describes all possible API parameters for FRED. Only
a select subset of them are allowed for a specific function, so see the
\code{Optional Parameter} section of each function to learn which ones are
allowed.
}
