% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_release_related_tags.R
\name{fredr_release_related_tags}
\alias{fredr_release_related_tags}
\title{Get the related FRED tags for one or more FRED tags within a release}
\usage{
fredr_release_related_tags(release_id = NULL, tag_names = NULL,
  exclude_tag_names = NULL, tag_group_id = NULL, search_text = NULL,
  limit = NULL, offset = NULL, order_by = NULL, sort_order = NULL,
  realtime_start = NULL, realtime_end = NULL)
}
\arguments{
\item{release_id}{An integer ID of the release. \emph{Required parameter.}}

\item{tag_names}{A string of tags to include in the response.  Multiple tags
can be delimited by a semicolon in a single string. \emph{Required parameter.}}

\item{exclude_tag_names}{A string of tags to \emph{not} include in the response.
Multiple tags can be delimited by a semicolon in a single string.}

\item{tag_group_id}{A string representing the tag group id to filter tags by
type.  No filtering by default.  Possible values include:
\itemize{
\item \code{"freq"} - Frequency
\item \code{"gen"} - General or Concept
\item \code{"geo"} - Geography
\item \code{"geot"} - Geography Type
\item \code{"rls"} - Release
\item \code{"seas"} - Seasonal Adjustment
\item \code{"src"} - Source
}}

\item{search_text}{A string to match text of tags.  No matching by default.}

\item{limit}{An integer limit on the maximum number of results to return.
Defaults to \code{1000}, the maximum.}

\item{offset}{An integer used in conjunction with \code{limit} for long series.
This mimics the idea of \emph{pagination} to retrieve large amounts of data over
multiple calls. Defaults to \code{0}.}

\item{order_by}{Order results by values of the specified attribute.
Possible values include: \code{"series_count"} (default), \code{"popularity"``,}"created"\code{,}"name"\code{,}"group_id"`.}

\item{sort_order}{A string representing the order of the resulting series.
Possible values are: \code{"asc"} (default), and \code{"desc"}.}

\item{realtime_start}{A \code{Date} indicating the start of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{realtime_end}{A \code{Date} indicating the end of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}
}
\value{
A \code{tibble} object.
}
\description{
Get the related FRED tags for one or more FRED tags within a release
}
\section{API Documentation}{


\href{https://research.stlouisfed.org/docs/api/fred/release_related_tags.html}{fred/release/related_tags}
}

\examples{
\dontrun{
fredr_release_related_tags(release_id = 10L, tag_names = "gnp")
}
}
\seealso{
\code{\link[=fredr_releases]{fredr_releases()}},  \code{\link[=fredr_releases_dates]{fredr_releases_dates()}}, \code{\link[=fredr_release]{fredr_release()}},
\code{\link[=fredr_release_dates]{fredr_release_dates()}}, \code{\link[=fredr_release_series]{fredr_release_series()}}, \code{\link[=fredr_release_sources]{fredr_release_sources()}},
\code{\link[=fredr_release_tags]{fredr_release_tags()}}, \code{\link[=fredr_release_tables]{fredr_release_tables()}}
}
