% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series.R
\name{fredr_series}
\alias{fredr_series}
\title{Return a FRED dataset as an \code{xts} object.}
\usage{
fredr_series(series_id, ...)
}
\arguments{
\item{series_id}{A string ID for the FRED series}

\item{...}{Additional parameters passed directly to API.  Must be of the
form \code{param_key = 'param_value'}.
required.}
}
\value{
An \code{xts} object.
}
\description{
Return a FRED dataset as an \code{xts} object.
}
\examples{
fredr_series(series_id = "UNRATE",
             observation_start = "1980-01-01",
             observation_end = "2000-01-01",
             unit = "chg")
fredr_series(series_id = "OILPRICE",
             frequency = "q",
             aggregation_method = "avg",
             unit = "log")
}
