% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series_search.R
\name{fredr_series_search}
\alias{fredr_series_search}
\alias{fredr_series_search_text}
\alias{fredr_series_search}
\alias{fredr_series_search_id}
\title{Search for a FRED series.}
\usage{
fredr_series_search_text(search_text = NULL, limit = 1000L, offset = 0,
  order_by = NULL, sort_order = "asc", filter_variable = NULL,
  filter_value = NULL, realtime_start = NULL, realtime_end = NULL,
  tag_names = NULL, exclude_tag_names = NULL)

fredr_series_search_id(search_text = NULL, limit = 1000L, offset = 0,
  order_by = NULL, sort_order = "asc", filter_variable = NULL,
  filter_value = NULL, realtime_start = NULL, realtime_end = NULL,
  tag_names = NULL, exclude_tag_names = NULL)
}
\arguments{
\item{search_text}{A string containing the words to match against economic
data series. For use with \code{\link{fredr_series_search_text}} and
\code{\link{fredr_series_search_id}}.}

\item{limit}{An integer limit on the maximum number of results to return.
Defaults to \code{1000}, the maximum.}

\item{offset}{An integer used in conjunction with \code{limit} for long series.
This mimics the idea of \emph{pagination} to retrieve large amounts of data over
multiple calls. Defaults to \code{0}.}

\item{order_by}{A string indicating the attribute to order results by.
Defaults to \code{"search_rank"} for \code{\link[=fredr_series_search_text]{fredr_series_search_text()}} and \code{"series_id"}
for \code{\link[=fredr_series_search_id]{fredr_series_search_id()}}. Possible values are:
\itemize{
\item \code{"search_rank"}
\item \code{"series_id"}
\item \code{"title"}
\item \code{"units"}
\item \code{"frequency"}
\item \code{"seasonal_adjustment"}
\item \code{"realtime_start"}
\item \code{"realtime_end"}
\item \code{"last_updated"}
\item \code{"observation_start"}
\item \code{"observation_end"}
\item \code{"popularity"}
\item \code{"group_popularity"}
}}

\item{sort_order}{A string representing the order of the resulting series.
Possible values are: \code{"asc"}, and \code{"desc"}. Defaults to \code{"asc"}.}

\item{filter_variable}{A string indicating the attribute to filter results
by. Possible values are: \code{"frequency"}, \code{"units"}, \code{"seasonal_adjustment"}.
Defaults to no filter.}

\item{filter_value}{The value of the \code{filter_variable} attribute to filter
by. Possible values depend on the value of \code{filter_variable}. Defaults to
no filter.}

\item{realtime_start}{A \code{Date} indicating the start of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{realtime_end}{A \code{Date} indicating the end of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{tag_names}{A semicolon delimited string of tag names that series match \emph{all} of.  Defaults to no tag filtering.}

\item{exclude_tag_names}{A semicolon delimited string of tag names that
series match \emph{none} of.  Defaults to no tag filtering.}
}
\value{
A \code{tibble} object.
}
\description{
Search FRED for a series by full text of series or by series ID.
}
\examples{
\dontrun{
library(fredr)
fredr_series_search_text("oil", order_by = "popularity", limit = 10)
fredr_series_search_text("oil", order_by = "search_rank", limit = 10, tag_names = "usa")
fredr_series_search_text("oil", filter_variable = "frequency", filter_value = "q")
fredr_series_search_id("UNRATE", order_by = "observation_start")
}
}
\references{
API Documentation:

\href{https://research.stlouisfed.org/docs/api/fred/series_search.html}{series/search}
}
\seealso{
\code{\link[=fredr_series_search_tags]{fredr_series_search_tags()}}, \code{\link[=fredr_series_search_related_tags]{fredr_series_search_related_tags()}}, \link{fredr_tags}
}
