% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series.R
\name{get_freq}
\alias{get_freq}
\title{Check the default frequency of a series with a call to the FRED API}
\usage{
get_freq(series_id)
}
\arguments{
\item{series_id}{A string}
}
\value{
An integer representing the number of series observations in a unit 
        of time.  Used in creating a \code{ts} object with the appropriate 
        temporal frequency.
}
\description{
Check the default frequency of a series with a call to the FRED API
}
\examples{
 get_freq(series_id = "GNPCA")
}

