% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FredR.R
\name{FredR}
\alias{FredR}
\title{FredR: R Interface to Federal Reserve Economic Data API}
\usage{
series.observations("12345")
}
\arguments{
\item{api.key}{API key (required). You can get the key by registering at the
following address: http://api.stlouisfed.org/api_key.html}

\item{category_id}{The FRED category ID}

\item{series_id}{A string}

\item{limit}{An integer between 1 and 100000}

\item{observation_start}{YYYY-MM-DD formatted string}

\item{observation_end}{YYYY-MM-DD formatted string}

\item{units}{One of the following values: 'lin', 'chg', 'ch1', '
pch', 'pc1', 'pca', 'cch', 'cca', 'log'}

\item{frequency}{One of the following values: 'd', 'w', 'bw', 'm', 'q', 'sa', 'a',
'wef', 'weth', 'wew', 'wetu', 'wem', 'wesu', 'wesa', 'bwew', 'bwem'}

\item{aggregation_method}{One of the following values: 'avg', 'sum', 'eop'}

\item{output_type}{One of the following values: '1', '2', '3', '4'}

\item{sort_order}{One of the following strings: 'asc', 'desc'.}

\item{offset}{non-negative integer, optional, default: 0}
}
\value{
List object of class FredR. The list object consists of functions
that perform the API calls (to be documented soon). For now, one can find
the descriptions of particular calls and their parameters here: http://api.stlouisfed.org/docs/fred/?utm_source=research&utm_medium=website&utm_campaign=data-tools

a datatable

A datatable
}
\description{
FredR: R Interface to Federal Reserve Economic Data API
}
\details{
This package provides an interface to the Federal Reserve Economic Data
(FRED) API. FRED covers 240,000 US and international macroeconomic time
series from 77 sources (including the World Bank,  OECD, and BIS).
FredR has been designed in a way that gives a user a complete interface to
all available functionalities of the API.
}
\author{
Janko Cizel

Janko Cizel
}
