% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr.R
\name{fredr}
\alias{fredr}
\title{Send a request to the FRED API}
\usage{
fredr(endpoint, ..., to_frame = TRUE, print_req = FALSE)
}
\arguments{
\item{endpoint}{A string representing the FRED API path of interest. See
\code{fredr_endpoints} for a list of endpoint possible values.}

\item{...}{A series of named parameters to be used in the query.  Must be of the form
\code{param_key = "param_value"}.  Acceptable parameters are endpoint-specific.
See  \code{\link{fredr_endpoints}} for a list of endpoints and \code{\link{fredr_docs}}
access to the endpoint web documentation.}

\item{to_frame}{A boolean value indicating whether or not the response
should be parsed and formatted as a data frame.  If \code{FALSE},
a \code{response} object is returned and further processing can be done with
\code{\link[httr]{content}}.  Default is \code{TRUE}.}

\item{print_req}{A boolean value indicating whether or not the request
should be printed as well.  Useful for debugging.  Default is \code{FALSE}.}
}
\value{
If \code{to_frame = TRUE}, a data frame containing the parsed response.
        If \code{to_frame = FALSE}, an object returned directly from \code{\link[httr]{GET}}.
}
\description{
Send a request to the FRED API
}
\examples{
 fredr(endpoint = "series/observations",
       series_id = "GNPCA",
       observation_start = "1990-01-01",
       observation_end = "2000-01-01")
 \dontrun{
 library(httr)
 library(tidyverse)
 resp <- fredr(endpoint = "series/observations",
               series_id = "GNPCA",
               observation_start = "1990-01-01",
               observation_end = "2000-01-01",
               to_frame = FALSE)
 data <- resp \%>\%
   httr::content() \%>\%
   .$observations \%>\%
   purrr::map_df(.f = function(x) tibble::as_tibble(x = x))
}
}
\references{
\url{https://api.stlouisfed.org/docs/fred/}.
}
