% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr.R
\name{fredr}
\alias{fredr}
\title{Send a request to the FRED API server}
\usage{
fredr(endpoint, ..., to_frame = TRUE, print_req = FALSE)
}
\arguments{
\item{endpoint}{A string representing the FRED API path of interest. See
https://api.stlouisfed.org/docs/fred/.}

\item{...}{A series of paramters to be used in the query.  Of the form
param_key = 'param_value'.  Acceptable parameters are specific to
the API path.  See https://api.stlouisfed.org/docs/fred/.}

\item{to_frame}{A boolean value indicating whether or not the response
should be parsed and formatted as a tbl_df.  Default is TRUE.}

\item{print_req}{A boolean value indicating whether or not the request
should be printed as well.  Useful for debugging.}
}
\value{
A tbl_df.  If to_frame is TRUE, return is a reponse object from 
       httr::GET().
}
\description{
Send a request to the FRED API server
}
\examples{
 fredr("series/observations",
       series_id = "GNPCA",
       observation_start = "1990-01-01",
       observation_end = "2000-01-01")
}

