% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_category.R
\name{fredr_category}
\alias{fredr_category}
\title{Get a category}
\usage{
fredr_category(category_id = 0L)
}
\arguments{
\item{category_id}{An integer ID for the category.  Default is \code{0} for the
root category. \emph{Required parameter.}}
}
\value{
A \code{tibble} object.
}
\description{
Get a category
}
\section{API Documentation}{


\href{https://research.stlouisfed.org/docs/api/fred/category.html}{fred/category}
}

\examples{
\dontrun{
fredr_category()
}
}
\seealso{
\code{\link[=fredr_category_children]{fredr_category_children()}}, \code{\link[=fredr_category_related]{fredr_category_related()}},
\code{\link[=fredr_category_series]{fredr_category_series()}}, \code{\link[=fredr_category_tags]{fredr_category_tags()}}, \code{\link[=fredr_category_related_tags]{fredr_category_related_tags()}}
}
