% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_docs.R
\name{fredr_docs}
\alias{fredr_docs}
\title{Open the web documentation for a certain FRED API topic.}
\usage{
fredr_docs(endpoint = "base", params = FALSE, debug = FALSE)
}
\arguments{
\item{endpoint}{A string representing the desired documentation for the exact
FRED API endpoint.  Default is 'base' and will open a link to
https://api.stlouisfed.org/docs/fred/.}

\item{params}{A boolean value.  If TRUE, the documentation will be opened at
the 'Paramters' section.  Default is FALSE.}

\item{debug}{A boolean value.  If TRUE, the documentation is not opened in a
browser.  Default is FALSE.}
}
\description{
Open the web documentation for a certain FRED API topic.
}
\examples{
\dontrun{
fredr_docs()
fredr_docs('category')
fredr_docs('series/observations')
fredr_docs('series/observations', params = TRUE)
}
}
