% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_endpoints.R
\docType{data}
\name{fredr_endpoints}
\alias{fredr_endpoints}
\title{List of available FRED API endpoints.}
\format{A data frame with 31 rows and 3 variables:
\describe{
  \item{type}{endpoint type (e.g. "Categories", "Releases", "Series", "Sources", and "Tags".)}
  \item{endpoint}{endpoint name (e.g. "fred/category", "fredr/series/observations", "fredr/tags").
  This name can be supplied to the \code{endpoint} parameter in \code{\link[fredr]{fredr_docs}} to
  open the \href{https://api.stlouisfed.org/docs/fred/}{FRED API endpoint documentation} in a web browser.}
  \item{note}{endpoint details}
}}
\source{
\url{https://api.stlouisfed.org/docs/fred/}.
}
\usage{
fredr_endpoints
}
\description{
List of available FRED API endpoints.
}
\seealso{
\code{\link[fredr]{fredr_docs}}
}
\keyword{datasets}
