% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_search.R
\name{fredr_search}
\alias{fredr_search}
\alias{fredr_search_id}
\alias{fredr_search_tags}
\alias{fredr_search_rel_tags}
\title{Search for a FRED series.}
\usage{
fredr_search(search = NULL, ...)

fredr_search_id(search = NULL, ...)

fredr_search_tags(search = NULL, ...)

fredr_search_rel_tags(search = NULL, tag_names = NULL, ...)
}
\arguments{
\item{search}{A string containing the search text.  Searches are performed against series
attributes (\code{\link{fredr_search}}, \code{\link{fredr_search_tags}}, and
\code{\link{fredr_search_rel_tags}}) or series ID
(\code{\link{fredr_search_id}}).  See Details for more information.}

\item{...}{A series of named parameters to be used in the query.  Must be of the form
\code{param_key = "param_value"}.  Acceptable parameters are endpoint-specific.
See  \code{\link{fredr_endpoints}} for a list of endpoints and \code{\link{fredr_docs}}
access to the endpoint web documentation.}

\item{tag_names}{A string.  Note that this value must be a valid FRED series tag
(e.g. \code{30-year}, \code{usa}, \code{quarterly}, etc.).}
}
\value{
A data frame.
}
\description{
Search FRED for a series.  Searches can be performed by the text of the series
(\code{fredr_search}), by series ID (\code{fredr_search_id}), by series tags
(\code{fredr_search_tags}), or related tags (\code{fredr_search_rel_tags}).
}
\details{
Search by various series characteristics:
\itemize{
  \item{\code{\link{fredr_search}}}{Get economic data series that match search text (by full text of series).}
  \item{\code{\link{fredr_search_id}}}{Get economic data series that match search text (by series ID).}
  \item{\code{\link{fredr_search_tags}}}{Get the FRED tags for a series search.}
  \item{\code{\link{fredr_search_rel_tags}}}{Get the related FRED tags for one or more FRED tags matching a series search.
  See below for more information about searching for related tags.}
}
From the FRED API documentation on \code{series/search/related_tags}: FRED tags are
attributes assigned to series. For this request, related FRED tags are the tags assigned
to series that match all tags in the \code{tag_names} parameter, no tags in the \code{exclude_tag_names}
parameter, and the search words set by the \code{search} parameter.
}
\examples{
\dontrun{
library(fredr)
fredr_search("oil", order_by = "popularity")
fredr_search_id("UNRATE", order_by = "popularity")
fredr_search_tags("gnp")
fredr_search_rel_tags("gnp", "usa")
}
}
\references{
See FRED API documentation on
\href{https://research.stlouisfed.org/docs/api/fred/series_search_related_tags.html}{searching related tags}.
}
