% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_search.R
\name{fredr_search}
\alias{fredr_search}
\alias{fredr_search_id}
\alias{fredr_search_tags}
\alias{fredr_search_rel_tags}
\title{Search for a FRED series.}
\usage{
fredr_search(search_text = NULL, limit = 100000L, offset = 0,
  order_by = NULL, sort_order = "asc", filter_variable = NULL,
  filter_value = NULL, realtime_start = NULL, realtime_end = NULL)

fredr_search_id(search_text = NULL, ...)

fredr_search_tags(series_search_text = NULL, ...)

fredr_search_rel_tags(series_search_text = NULL, tag_names = NULL, ...)
}
\arguments{
\item{search_text}{A string containing the words to match against economic
data series. For use with \code{\link{fredr_search}} and
\code{\link{fredr_search_id}}. See Details for more information.}

\item{order_by}{A string indicating the attribute to order results by.
Defaults to \code{"search_rank"} if \code{search_type = "full_text"} and \code{"series_id"}
if \code{search_type = "series_id"}. Possible values are:
\itemize{
\item \code{"search_rank"}
\item \code{"series_id"}
\item \code{"title"}
\item \code{"units"}
\item \code{"frequency"}
\item \code{"seasonal_adjustment"}
\item \code{"realtime_start"}
\item \code{"realtime_end"}
\item \code{"last_updated"}
\item \code{"observation_start"}
\item \code{"observation_end"}
\item \code{"popularity"}
\item \code{"group_popularity"}
}}

\item{filter_variable}{A string indicating the attribute to filter results
by. Possible values are: \code{"frequency"}, \code{"units"}, \code{"seasonal_adjustment"}.
Defaults to no filter.}

\item{filter_value}{The value of the \code{filter_variable} attribute to filter
by. Possible values depend on the value of \code{filter_variable}. Defaults to
no filter.}

\item{series_search_text}{A string containing the search text.  For use with \code{\link{fredr_search_tags}}
and \code{\link{fredr_search_rel_tags}}.  See Details for more information.}

\item{exclude_tag_names}{A semicolon delimited string of tag names that
series match \emph{none} of.}
}
\value{
A data frame.
}
\description{
Search FRED for a series.  Searches can be performed by the text of the series
(\code{fredr_search}), by series ID (\code{fredr_search_id}), by series tags
(\code{fredr_search_tags}), or related tags (\code{fredr_search_rel_tags}).
}
\details{
Search by various series characteristics:

\itemize{
\item{\code{\link{fredr_search}}}{Get economic data series that match search text (by full text of series).}
\item{\code{\link{fredr_search_id}}}{Get economic data series that match search text (by series ID).}
\item{\code{\link{fredr_search_tags}}}{Get the FRED tags for a series search.}
\item{\code{\link{fredr_search_rel_tags}}}{Get the related FRED tags for one or more FRED tags matching a series search.
See below for more information about searching for related tags.}
}
From the FRED API documentation on \code{series/search/related_tags}: FRED tags are
attributes assigned to series. For this request, related FRED tags are the tags assigned
to series that match all tags in the \code{tag_names} parameter, no tags in the \code{exclude_tag_names}
parameter, and the search words set by the \code{series_search_text} parameter.
}
\examples{
\dontrun{
library(fredr)
fredr_search("oil", order_by = "popularity")
fredr_search_id("UNRATE", order_by = "popularity")
fredr_search_tags("gnp")
fredr_search_rel_tags("gnp", "usa")
}
}
\references{
See FRED API documentation on
\href{https://research.stlouisfed.org/docs/api/fred/series_search_related_tags.html}{searching related tags}.
}
\seealso{
\code{\link{fredr_tags}}
}
