% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series.R
\name{fredr_series}
\alias{fredr_series}
\title{Return a FRED series as an \code{xts} object.}
\usage{
fredr_series(series_id, ...)
}
\arguments{
\item{series_id}{A string ID for the FRED series.  Use \code{\link{fredr_search}}
to search for \code{series_id} of a series.}

\item{...}{Additional parameters passed directly to FRED API.  Must be of the
form \code{param_key = 'param_value'}.}
}
\value{
An \code{xts} object with observation dates in the index and parsed numeric values.
}
\description{
Return a FRED series as an \code{xts} object.
}
\examples{
fredr_series(series_id = "UNRATE",
             observation_start = "1980-01-01",
             observation_end = "2000-01-01",
             unit = "chg")
fredr_series(series_id = "OILPRICE",
             frequency = "q",
             aggregation_method = "avg",
             unit = "log")
}
