% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series.R
\name{fredr_series}
\alias{fredr_series}
\title{Fetch a FRED series}
\usage{
fredr_series(series_id, options = fredr_options())
}
\arguments{
\item{series_id}{A string ID for the FRED series. Use \code{\link[=fredr_search]{fredr_search()}}
to find the \code{series_id} of a series.}

\item{options}{Extra options defined by \code{\link[=fredr_options]{fredr_options()}}.}
}
\value{
A \code{tibble} object with observation dates and values.
}
\description{
Given a series ID, return observations of that series as a \code{tibble} object.
}
\section{Optional Parameters}{

\itemize{
\item \code{observation_start}
\item \code{observation_end}
\item \code{frequency}
\item \code{aggregation_method}
\item \code{limit}
\item \code{offset}
\item \code{sort_order}
\item \code{units}
\item \code{realtime_start}
\item \code{realtime_end}
\item \code{vintage_dates}
\item \code{output_type}
}
}

\section{API Documentation}{


\href{https://research.stlouisfed.org/docs/api/fred/series_observations.html}{fred/series/observations}
}

\examples{
fredr_series(
  series_id = "UNRATE",
  options = fredr_options(
    observation_start = as.Date("1980-01-01"),
    observation_end = as.Date("2000-01-01"),
    unit = "chg"
  )
)

fredr_series(
  series_id = "OILPRICE",
  options = fredr_options(
    frequency = "q",
    aggregation_method = "avg",
    unit = "log"
  )
)

}
