% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series.R
\name{fredr_series}
\alias{fredr_series}
\title{Return a FRED series as an \code{xts} object.}
\usage{
fredr_series(series_id, ...)
}
\arguments{
\item{series_id}{A string ID for the FRED series.  Use \code{\link{fredr_search}}
to find the \code{series_id} of a series.}

\item{...}{A series of named parameters to be used in the query.  Must be of the form
\code{param_key = "param_value"}.  Acceptable parameters are endpoint-specific.
See  \code{\link{fredr_endpoints}} for a list of endpoints and \code{\link{fredr_docs}}
access to the endpoint web documentation.}
}
\value{
An \code{xts} object with observation dates in the index corresponding to
parsed numeric observations as values.
}
\description{
Given a series ID, return observations of that series as an \code{xts} object.
}
\examples{
fredr_series(series_id = "UNRATE",
             observation_start = "1980-01-01",
             observation_end = "2000-01-01",
             unit = "chg")
fredr_series(series_id = "OILPRICE",
             frequency = "q",
             aggregation_method = "avg",
             unit = "log")
}
