% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_series.R
\name{fredr_series}
\alias{fredr_series}
\title{Return a FRED dataset as an \code{xts} object.}
\usage{
fredr_series(...)
}
\arguments{
\item{...}{A series of paramters to be used in the query.  Of the form
param_key = 'param_value'.  The parameter \code{series_id} is
required.}
}
\value{
An \code{xts} object.
}
\description{
Return a FRED dataset as an \code{xts} object.
}
\examples{
fredr_series(series_id = "UNRATE",
             observation_start = "1980-01-01",
             observation_end = "2000-01-01",
             unit = "chg")
fredr_series(series_id = "OILPRICE",
             frequency = "q",
             aggregation_method = "avg",
             unit = "log")
}
