% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_set_key.R
\name{fredr_set_key}
\alias{fredr_set_key}
\title{Set the FRED API key}
\usage{
fredr_set_key(key)
}
\arguments{
\item{key}{A FRED API key as a string.  Obtain one at
\url{https://api.stlouisfed.org/api_key.html}.}
}
\description{
Users \code{fredr} package must authenticate with the FRED API by use of an API key.
The function \code{fredr_key} sets the FRED API key as an environment variable for
use with the service.
}
\details{
\code{fredr_set_key} sets a key as an environment variable for use with
the \code{fredr} package in the current session.  The key can also be set in the
\code{.Renviron} file in the working directory.  You can edit the file manually by
appending the line \code{FRED_API_KEY = my_api_ky}, where \code{my_api_key} is your
actual key.  The function `usethis::edit_r_environ()` does this safely.  Run
\code{readRenviron(".Renviron")} to set the key in the current session; the
variable will be set in subsequent sessions started in the working directory.
}
\examples{

\dontrun{
library(fredr)
fredr_set_key("abcdefghijklmnopqrstuvwxyz123456")
}

}
\references{
See St. Louis Fed Web Services \href{https://api.stlouisfed.org/api_key.html}{API Keys} to obtain an API key.
}
\seealso{
Note that by using a FRED API key, you agree to the FRED API
\href{https://research.stlouisfed.org/docs/api/terms_of_use.html}{Terms of Use}.
}
