% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_tags.R
\name{fredr_tags}
\alias{fredr_tags}
\title{Get FRED series tags}
\usage{
fredr_tags(realtime_start = NULL, realtime_end = NULL, tag_names = NULL,
  tag_group_id = NULL, search_text = NULL, limit = NULL, offset = NULL,
  order_by = NULL, sort_order = NULL)
}
\arguments{
\item{realtime_start}{A \code{Date} indicating the start of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{realtime_end}{A \code{Date} indicating the end of the real-time period.
Defaults to today's date. For more information, see
\href{https://research.stlouisfed.org/docs/api/fred/realtime_period.html}{Real-Time Periods}.}

\item{tag_names}{A semicolon delimited string of tag names to only
include in the response. No filtering by tag names by default.}

\item{tag_group_id}{A string tag group id to filter tags by type. No filtering
by tag group by default. Possible values are:
\itemize{
\item \code{"freq"} = Frequency
\item \code{"gen"} = General or Concept
\item \code{"geo"} = Geography
\item \code{"geot"} = Geography Type
\item \code{"rls"} = Release
\item \code{"seas"} = Seasonal Adjustment
\item \code{"src"} = Source
}}

\item{search_text}{A string indicating the words to find matching tags with.
No filtering by search words by default.}

\item{limit}{An integer limit on the maximum number of results to return.
Defaults to \code{1000}, the maximum.}

\item{offset}{An integer used in conjunction with \code{limit} for long series.
This mimics the idea of \emph{pagination} to retrieve large amounts of data over
multiple calls. Defaults to \code{0}.}

\item{order_by}{Order results by values of the specified attribute. Default
ordering is by \code{series_count}. Possible values are:
\itemize{
\item \code{"series_count"}
\item \code{"popularity"}
\item \code{"created"}
\item \code{"name"}
\item \code{"group_id"}
}}

\item{sort_order}{A string representing the order of the resulting series,
sorted by the attribute values specified by \code{order_by}. Possible values
are: \code{"asc"}, and \code{"desc"}. Defaults to \code{"asc"}.}
}
\value{
A data frame containing tags and their descriptions.
}
\description{
Get FRED tags. Optionally, filter results by tag name, tag group, or
search text. FRED tags are attributes assigned to a series. By default,
all tags are returned, unfiltered, up to the \code{limit}.
}
\section{API Documentation}{


\href{https://research.stlouisfed.org/docs/api/fred/tags.html}{fred/tags}
}

\examples{
\dontrun{
fredr_tags()
fredr_tags(tag_names = "gdp;oecd")
fredr_tags(tag_group_id = "geo")
fredr_tags(search_text = "unemployment")
}
}
\seealso{
\code{\link[=fredr_docs]{fredr_docs()}}, \code{\link[=fredr_search]{fredr_search()}}, \code{\link[=fredr_series]{fredr_series()}}
}
