% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredr_tags.R
\name{fredr_tags}
\alias{fredr_tags}
\title{Get FRED series tags}
\usage{
fredr_tags(...)
}
\arguments{
\item{...}{A series of named parameters to be used in the query.  Must be of the form
\code{param_key = "param_value"}.  See the \href{https://research.stlouisfed.org/docs/api/fred/tags.html}{Tags}
documentation for a list of accepted parameters.}
}
\value{
a data frame.
}
\description{
List or search (by series text, tag names, or by tag group id) FRED series tags.
}
\examples{
\dontrun{
library(tidyverse)
fredr_tags()
fredr_tags(tag_names = "gdp;oecd")
fredr_tags(tag_group_id = "geo")
fredr_tags(search_text = "unemployment")
}
}
\references{
\url{https://research.stlouisfed.org/docs/api/fred/tags.html}.
}
\seealso{
\code{\link[fredr]{fredr_docs}}, \code{\link[fredr]{fredr_search}}, \code{\link[fredr]{fredr_series}}
}
